scripts/bootstrap/requirements-kafka.txt
confluent-kafka==2.8.2
PyYAML==6.0.2
requests==2.32.3
authlib>=1.3.0
httpx>=0.25.0